+++ src/interfaces/IChildERC721.sol
@@ -6,7 +6,7 @@ interface IChildERC721 {
 
     function bridge() external view returns (address);
 
-    function parent() external view returns (address token, uint16 chainId);
+    function parent() external view returns (address token, uint64 chainId);
 
     function mint(address to, uint tokenId) external;
